マスターと穴埋のスライドのスクリプトを統合

+++ pandoc/script/lecture_beamer.sh
@@ -11,32 +11,27 @@ BASENAME=${FILENAME%.*}
 
 cp $FILENAME $BASENAME-temp.md
 
-mkdir -p temp
-cp plots_figures/*.* temp/
+if [ -d plots_figures ]; then
+    mkdir -p temp
+    cp plots_figures/*.* temp/
 
-# 図・グラフの参照先フォルダをtempに変更
-sed -i 's/plots_figures\//temp\//g' $BASENAME-temp.md
+    # 図・グラフの参照先フォルダをtempに変更
+    sed -i 's/plots_figures\//temp\//g' $BASENAME-temp.md
 
-# texの図などの強調を太字から空欄または赤字に変更
-if [ "$BLANKMODE" = "TRUE" ]; then
-  sed -i  's/\\textbf/\\phantom/g' temp/*.tex
-else
-  sed -i 's/\\textbf/\\textcolor{red}/g' temp/*.tex
-fi
-
-# gnuplotの空欄スライド用描画コマンドを削除
-sed -i '/\#plot-blank\#/{n;d;}' temp/*.plt
-
-
-# -bが指定された場合，tikzの赤字・青字や塗りを透明に
-if [ "$BLANKMODE" = "TRUE" ]; then
-sed -i   \
-            -e 's/\[red/\[transparent/g' \
-            -e 's/text=red/text=white/g' \
-            -e 's/\[blue/\[transparent/g' \
-            -e 's/text=blue/text=white/g'  \
-            -e 's/opacity=0.1/opacity=0/g' \
-        temp/*.tex
+    # -bが指定された場合，tikzの赤字・青字や塗りを透明に
+    if [ "$BLANKMODE" = "TRUE" ]; then
+      sed -i  's/\\textbf/\\phantom/g' temp/*.tex  # texの図などの強調を太字から空欄に変更
+      sed -i '/\#plot-master\#/{n;d;}' temp/*.plt  # gnuplotのマスタースライド用描画コマンドを削除
+      sed -i -e 's/\[red/\[transparent/g' \
+             -e 's/text=red/text=white/g' \
+             -e 's/\[blue/\[transparent/g' \
+             -e 's/text=blue/text=white/g'  \
+             -e 's/opacity=0.1/opacity=0/g' \
+            temp/*.tex
+    else
+      sed -i 's/\\textbf/\\textcolor{red}/g' temp/*.tex  # texの図などの強調を太字から赤字に変更
+      sed -i '/\#plot-blank\#/{n;d;}' temp/*.plt  # gnuplotの空欄スライド用描画コマンドを削除
+    fi
 fi
 
 # Pandocでマークダウンをtexに変換
@@ -70,12 +65,21 @@ fi
 if [ "$DEBUGMODE" = "TRUE" ]; then
   exit 0
 fi
+
 rm $BASENAME-temp.md
-rm $BASENAME-gnuplottex-fig*.gnuplot
-rm $BASENAME-gnuplottex-fig*.tex
-rm $BASENAME.gnuploterrors
 rm $BASENAME.tex
+
+if [ ! -d temp ]; then
+  exit 0
+fi
+
 rm -r temp
+
+if [ -e $BASENAME.gnuploterrors ]; then
+  rm $BASENAME.gnuploterrors
+  rm $BASENAME-gnuplottex-fig*.gnuplot
+  rm $BASENAME-gnuplottex-fig*.tex
+fi
